+++ README.md
@@ -9,4 +9,4 @@
 ---
 
 - [x] 전체 상품 조회에 페이지네이션 적용하기
-- [ ] 위시리스트 조회에 페이지네이션 적용하기
+- [x] 위시리스트 조회에 페이지네이션 적용하기
