+++ src/main/java/gift/service/WishService.java
@@ -11,7 +11,7 @@ public interface WishService {
 
     WishIdResponseDto addProduct(WishAddRequestDto wishAddRequestDto, String email);
 
-    List<WishResponseDto> getWishList(String email);
+    List<WishResponseDto> getWishList(String email, int pageNo, String sortBy);
 
     void deleteProduct(String email, Long wishId, WishDeleteRequestDto wishDeleteRequestDto);
