+++ src/test/java/gift/jpaTest/MemberTest.java
@@ -2,8 +2,8 @@ package gift.jpaTest;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import gift.constant.MemberRole;
 import gift.entity.Member;
-import gift.entity.MemberRole;
 import gift.repository.MemberRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
