+++ src/test/java/gift/jpaTest/WishTest.java
@@ -3,8 +3,8 @@ package gift.jpaTest;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import gift.constant.MemberRole;
 import gift.entity.Member;
-import gift.entity.MemberRole;
 import gift.entity.Product;
 import gift.entity.Wish;
 import gift.repository.MemberRepository;
